--- src/pages/shared/Team/Members/index.js
@@ -1,1 +1,0 @@
-export { Members } from './Members'
