+++ src/pages/shared/Team/TeamMembers/index.js
@@ -1,0 +1,1 @@
+export { TeamMembers } from './TeamMembers'
